+++ Lombiq.Hosting.Tenants.Admin.Login/Constants/FeatureNames.cs
@@ -1,0 +1,9 @@
+namespace Lombiq.Hosting.Tenants.Admin.Login.Constants
+{
+    public static class FeatureNames
+    {
+        public const string Module = "Lombiq.Hosting.Tenants.Admin.Login";
+
+        public const string SubTenant = Module + "." + nameof(SubTenant);
+    }
+}
