+++ Lombiq.Hosting.Tenants.Admin.Login/Manifest.cs
@@ -1,8 +1,8 @@
 using OrchardCore.Modules.Manifest;
-using static Lombiq.Hosting.Tenants.Admin.Constants.FeatureNames;
+using static Lombiq.Hosting.Tenants.Admin.Login.Constants.FeatureNames;
 
 [assembly: Module(
-    Name = "Lombiq Hosting - Tenants Admin",
+    Name = "Lombiq Hosting - Tenants Admin Login ",
     Author = "Lombiq Technologies",
     Website = "https://github.com/Lombiq/Hosting-Tenants",
     Version = "1.0.0"
@@ -10,7 +10,7 @@ using static Lombiq.Hosting.Tenants.Admin.Constants.FeatureNames;
 
 [assembly: Feature(
     Id = Module,
-    Name = "Lombiq Hosting - Tenants Admin",
+    Name = "Lombiq Hosting - Tenants Admin Login",
     Description = "Ability to log in as a tenant’s admin user.",
     Category = "Hosting",
     DefaultTenantOnly = true,
@@ -19,7 +19,7 @@ using static Lombiq.Hosting.Tenants.Admin.Constants.FeatureNames;
 
 [assembly: Feature(
     Id = SubTenant,
-    Name = "Lombiq Hosting - Tenants Admin - Sub-tenant",
+    Name = "Lombiq Hosting - Tenants Admin Login - Sub-tenant",
     Description = "Adds the ability to log in to the tenant from the Default tenant.",
     Category = "Hosting",
     IsAlwaysEnabled = true
