+++ Lombiq.Hosting.Tenants.Admin.Login/Startup.cs
@@ -1,13 +1,13 @@
-using Lombiq.Hosting.Tenants.Admin.Constants;
-using Lombiq.Hosting.Tenants.Admin.Filters;
-using Lombiq.Hosting.Tenants.Admin.Permissions;
-using Lombiq.Hosting.Tenants.Admin.Services;
+using Lombiq.Hosting.Tenants.Admin.Login.Constants;
+using Lombiq.Hosting.Tenants.Admin.Login.Filters;
+using Lombiq.Hosting.Tenants.Admin.Login.Permissions;
+using Lombiq.Hosting.Tenants.Admin.Login.Services;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.DependencyInjection;
 using OrchardCore.Modules;
 using OrchardCore.Security.Permissions;
 
-namespace Lombiq.Hosting.Tenants.Admin
+namespace Lombiq.Hosting.Tenants.Admin.Login
 {
     [Feature(FeatureNames.Module)]
     public class Startup : StartupBase
