--- Lombiq.Hosting.Tenants.Admin/Constants/FeatureNames.cs
@@ -1,9 +1,0 @@
-namespace Lombiq.Hosting.Tenants.Admin.Constants
-{
-    public static class FeatureNames
-    {
-        public const string Module = "Lombiq.Hosting.Tenants.Admin";
-
-        public const string SubTenant = Module + "." + nameof(SubTenant);
-    }
-}
